Reimplement path-checking using new methodology
Replaces old code for handling when a node is reached for adding it to the stack and choosing the next direction.
Still need to finish implementing repeated-node code; primary concern is when nodePrevious has a shorter path to start from current node and how far to update the dependencies. Ideally each previously explored path from nodePrevious back should be checked to see if the newly discovered route is shorter and update their data accordingly. I think a recursive function would actually be easiest to implement, but am not sure how to go about it or if our arduino has the power to execute such a greedy process; any recommendations would be helpful.