Upload smoother proposal 1
First actual smoothing algorithm proposal for the physical micromouse using template from smootherV0.c. Major tweaking of control variables is likely necessary to optimize the smoothing operation.
In testing, the code appears to be very resource intensive due to the indefinite iterative process used, so other algorithms should still be explored as possibilities to reduce processing time and improve overall performance. A different way of displaying the graph is also needed to make it easier to comprehend (grid size is too big for text-based output in simulator).