Upload smoother proposal 2
This proposal combines the basic diagonalization from smootherV0() with the intensive smoothing process from smootherV1(), with no major edits to either algorithm beyond passing smootherV0()'s algorithm to smootherV1()'s (this is a standalone function that doesn't need either smootherV0.c or smootherV1.c to be implemented). A better method of displaying the processed data is still necessary.