Reduce granularity in smootherV2.c
After testing US sensors; measuring down to .5cm would likely result in constant errors of being off by 1 unit, so the grid size has been halved to double the size of the units to 1cm. Also changed individual fixed values that were based on the grid ratios to equations relating to that ratio so that it can be easily changed only by editing the RATIO variable.